toolkit/content/tests/widgets/test_videocontrols_standalone.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Video controls test</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>

<pre id="test">
<script class="testbody" type="text/javascript">

const videoWidth = 320;
const videoHeight = 240;

function getMediaElement(aWindow) {
  return aWindow.document.getElementsByTagName("video")[0];
}

var popup = window.open("seek_with_sound.ogg");
popup.addEventListener("loadedmetadata", function onLoadedMetaData() {
  popup.removeEventListener("loadedmetadata", onLoadedMetaData);
  var video = getMediaElement(popup);
  runTestVideo(video);
});

function runTestVideo(aVideo) {
  var boundingRect = aVideo.getBoundingClientRect();
  is(boundingRect.width, videoWidth, "Width of the video should match expectation");
  is(boundingRect.height, videoHeight, "Height of video should match expectation");
  popup.close();
  runTestAudioPre();
}

function runTestAudioPre() {
  popup = window.open("audio.ogg");
  popup.addEventListener("loadedmetadata", function onLoadedMetaData() {
    popup.removeEventListener("loadedmetadata", onLoadedMetaData);
    var audio = getMediaElement(popup);
    runTestAudio(audio);
  });
}

function runTestAudio(aAudio) {
  var boundingRect = aAudio.getBoundingClientRect();
  is(boundingRect.height, 28, "Height of audio element should be 28, which is equal to the controls bar.");
  popup.close();
  SimpleTest.finish();
}

SimpleTest.waitForExplicitFinish();

</script>
</pre>
</body>
</html>
